+++ packages/rsocket-router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @powersync/service-rsocket-router
 
+## 0.2.13
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.8.1
+
 ## 0.2.12
 
 ### Patch Changes