+++ packages/service-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @powersync/service-client
 
+## 0.0.2
+
+### Patch Changes
+
+- 34b8561: Trusted publishing release
+
 ## 0.0.1
 
 ### Patch Changes